Rewrite of Opaque text output
I reworked the opaque text for ILI fonts...
to speed it up.

The code uses a lot of the earlier code that used
the drawfontbits.  The main difficulty is to figure
out the ares to fill that are above, below, to left and right of the
actual drawn character.

I think it is working pretty well.

The update also includes an update to the font output font program that
allows you to time the different types of fonts and see them at
different orientations.