ILI and GFX font centering code
Add centering of text for GFX and ILI fonts...

This includes both the setCursor(x, y, true);
format which when you do a text output calculates the width and height
of your output and then centers the text at the x, y coordinate.

Also supports the X and/or Y coordinate of CENTER which like above
computes the width and height of the text output and then centers on one
or both axis.

Added new example text4 which tries out these capabilities.

While doing this I found there were several issues with the setCursor
and getCursor functions that were inconsistent with the usage of
swapping the X and Y coordinates when in portrait mode.

The test sketch outputs characters (0, 1, 2, 3) near the 4 corners when it changes rotation, so you can see if the set cursor is working properly.